--- src/app/page.module.scss
@@ -1,82 +1,0 @@
-@use '@/shared/styles' as *;
-
-.container {
-  @include Flex;
-  height: 100%;
-
-  .title {
-    @include Title ($size: 40px);
-    margin-top: 24px;
-    margin-bottom: 100px;
-
-    >span {
-      color: $purple;
-    }
-  }
-
-
-
-  >h2 {
-    margin: 15px 0;
-
-    >span {
-      color: $logo-yellow;
-    }
-  }
-
-  >a>button {
-    width: 280px;
-  }
-
-}
-
-.logoWrapper {
-  position: relative;
-
-}
-
-.logoStart {
-  animation:
-    pulse-async 3s ease-in-out forwards,
-    expand 1.5s 3s ease-out forwards;
-}
-
-.logo {
-  margin-top: 30px;
-  animation:
-    rotate 3s ease infinite,
-}
-
-
-
-// remove after real TELEGRAM AUTH
-.columnWrapper {
-  @include Flex;
-  gap: 20px;
-
-  .tip {
-    width: 200px;
-    font-size: 14px;
-    color: $grey;
-    text-align: center;
-
-    >span {
-      font-size: 18px;
-      font-weight: 600;
-    }
-
-    >i {
-      font-size: 12px;
-    }
-  }
-
-  >button {
-    margin-bottom: 40px;
-  }
-}
-
-.initDataRaw {
-  display: flex;
-  padding-top: 5%;
-  gap: 20px;
-}
