--- src/components/AuthForm/AuthForm.tsx
@@ -1,32 +1,0 @@
-import { useForm, SubmitHandler } from 'react-hook-form';
-import { Button, Input } from '@/shared/ui';
-import { setAuthToken } from '@/shared/utils/auth';
-import styles from './authForm.module.scss';
-
-type AuthFormProps = {
-  onAuthSuccess: (id: string) => void;
-};
-
-type FormInputs = {
-  auth_id_input: number;
-};
-
-export function AuthForm({ onAuthSuccess }: AuthFormProps) {
-  const { register, handleSubmit, reset } = useForm<FormInputs>({
-    mode: 'onSubmit',
-  });
-
-  const onSubmit: SubmitHandler<FormInputs> = async (data) => {
-    const authId = String(data.auth_id_input);
-    setAuthToken(authId);
-    onAuthSuccess(authId);
-    reset();
-  };
-
-  return (
-    <form className={styles.formId} onSubmit={handleSubmit(onSubmit)}>
-      <Input placeholder={'type your id'} type={'number'} {...register('auth_id_input')} />
-      <Button type='submit' text={'Submit'} color='default' />
-    </form>
-  );
-}
