--- src/components/AuthForm/authForm.module.scss
@@ -1,18 +1,0 @@
-.formId {
-    width: 100%;
-    display: flex;
-    gap: 20px;
-
-    > :nth-child(1) {
-        width: 200px;
-
-        >input {
-            width: 100%;
-        }
-    }
-
-    >button {
-        width: 120px;
-        height: 50px;
-    }
-}
