--- src/components/index.ts
@@ -1,18 +1,0 @@
-//export { ChallengeInfo } from '../widgets/ChallengeInfo/ChallengeInfo';
-// export { ProgressBar } from '../features/ProgressBar/ProgressBar';
-// export { ModalDelete } from '../features/Modals/Modals';
-export { AuthForm } from './AuthForm/AuthForm';
-
-// export { Footer } from '../widgets/Footer/Footer';
-// export { Navigation } from '../widgets/Navigation/Navigation';
-// export { ListChallenges } from '../widgets/ListChallenges/ListChallenges';
-// export { CreateForm } from '../widgets/CreateForm/CreateForm';
-// export { EditForm } from '../widgets/EditForm/EditForm';
-//UI
-// export { Button } from '../shared/ui/Button/Button';
-// export { Input } from '../shared/ui/Input/Input';
-// export { Switcher } from '../shared/ui/Switcher/Switcher';
-
-// export { ErrorBoundary } from '../shared/ui/Errors/ErrorBoundary';
-// export { ErrorPage } from '../shared/ui/Errors/ErrorPage';
-// export { CardSkeleton } from '../shared/ui/Skeletons/Skeletons';
