+++ src/widgets/index.ts
@@ -4,3 +4,4 @@ export { InProgressChallengesList } from './InProgressChallengesList/InProgressC
 export { CreateForm } from './CreateForm/CreateForm';
 export { EditForm } from './EditForm/EditForm';
 export { ChallengeInfo } from './ChallengeInfo/ChallengeInfo';
+export { Greeting } from './Greeting/Greeting';
